requirements.txt
# FastAPI + Web
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
httpx==0.27.2
pydantic-settings==2.4.0
websockets>=12.0

# Pipecat orchestrator
pipecat-ai[silero,websocket]

# STT
faster-whisper==1.0.3

# TTS (CosyVoice 2 installed separately via git clone)

# Audio processing
soundfile==0.12.1
numpy>=1.26,<2.0

# HuggingFace
huggingface_hub

# Process management
supervisor
